requirements.txt
discord.py
numpy
Pillow
jproperties
opencv-python